refactor: changes AppContext into TasksContext
