+++ src/components/Header/index.tsx
@@ -1,12 +1,12 @@
 import { Button } from "@mui/material"
 import { useState } from "react"
-import { useApp } from "../../contexts/AppContext"
+import { useTasks } from "../../contexts/TasksContext"
 import DoneTasksModal from "../DoneTaskModal"
 import styles from "./Header.module.scss"
 
 const Header = () => {
     // States and Hooks
-    const { tasksList } = useApp()
+    const { tasksList } = useTasks()
     const showDoneTasksButton: boolean = tasksList.length > 0
     const [isDoneTasksModalOpen, setIsDoneTaskModalOpen] = useState(false)
