+++ src/contexts/TasksContext.tsx
@@ -17,7 +17,7 @@ export type Task = {
     status: Status
 }
 
-interface AppContextInterface {
+interface TasksContextInterface {
     tasksList: Task[]
     setTasksList(list: Task[]): void
     addNewTask(
@@ -59,19 +59,19 @@ const initialContextValue = {
     editTask: () => undefined,
 }
 
-const AppContext = createContext<AppContextInterface>(initialContextValue)
+const TasksContext = createContext<TasksContextInterface>(initialContextValue)
 
 // It's better to export a hook and use that instead of using the Context itself in other files.
 // (It's one of them best practices)
-export const useApp = () => {
+export const useTasks = () => {
     // One of the best practices about context is to make sure that the hook is being called in the right place
     // and not someplace random in the application that makes debugging harder.
-    const context = useContext(AppContext)
-    if (!context) throw new Error("useApp must be used within a AppContext")
+    const context = useContext(TasksContext)
+    if (!context) throw new Error("useTasks must be used within a TasksContext")
     return context
 }
 
-const AppProvider: React.FC<Props> = ({ children }) => {
+const TasksProvider: React.FC<Props> = ({ children }) => {
     // States and Hooks
     const [tasksList, setTasksList] = useState<Task[]>([])
 
@@ -152,7 +152,9 @@ const AppProvider: React.FC<Props> = ({ children }) => {
     }
 
     // Render
-    return <AppContext.Provider value={value}>{children}</AppContext.Provider>
+    return (
+        <TasksContext.Provider value={value}>{children}</TasksContext.Provider>
+    )
 }
 
-export default AppProvider
+export default TasksProvider
